Split facebook tests in multiple files.
